Initial screenshot test
New Screenshot Scene added. Screenshotter script can grab a screenshot of the viewport and save to an image file. Later this data will be used to make a web call to Azure Vision service, but this is a good start for this.

Also TextMeshPro was added to integrate with MRTK 3.